Lyft-Dental/virtual_dental_clinc/vdc/requirements.txt
Django==3.0.6
django-allauth==0.41.0
django-crispy-forms==1.9.1
django-filter==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
Pillow==5.1.0